+++ pkg/admission/admission.go
@@ -28,6 +28,7 @@ import (
 	"github.com/k8tz/k8tz/pkg/inject"
 	"github.com/k8tz/k8tz/pkg/version"
 	admission "k8s.io/api/admission/v1beta1"
+	batchv1 "k8s.io/api/batch/v1"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
@@ -43,6 +44,7 @@ type RequestsHandler struct {
 	InjectByDefault          bool
 	HostPathPrefix           string
 	LocalTimePath            string
+	CronJobTimeZone          bool
 	clientset                kubernetes.Interface
 }
 
@@ -54,6 +56,7 @@ func NewRequestsHandler() RequestsHandler {
 		InjectByDefault:          true,
 		HostPathPrefix:           inject.DefaultHostPathPrefix,
 		LocalTimePath:            inject.DefaultLocalTimePath,
+		CronJobTimeZone:          false,
 	}
 }
 
@@ -139,20 +142,20 @@ func (h *RequestsHandler) handleFunc(w http.ResponseWriter, r *http.Request) {
 }
 
 func (h *RequestsHandler) handleAdmissionReview(review *admission.AdmissionReview) (k8tz.Patches, error) {
-	if review.Request.Resource != podResource || review.Request.Operation != "CREATE" {
-		return nil, nil
-	}
-
-	var patches k8tz.Patches
-	var err error
-	if review.Request.Namespace != metav1.NamespacePublic && review.Request.Namespace != metav1.NamespaceSystem {
-		patches, err = h.handleAdmissionRequest(review.Request)
-		if err != nil {
-			return nil, err
+	if review.Request.Operation == "CREATE" && review.Request.Namespace != metav1.NamespaceSystem {
+		var patches k8tz.Patches
+		var err error
+		if review.Request.Resource == podResource {
+			patches, err = h.handlePodAdmissionRequest(review.Request)
+		} else if review.Request.Resource == cronJobResource {
+			patches, err = h.handleCronJobAdmissionRequest(review.Request)
 		}
+
+		return patches, err
 	}
 
-	return patches, nil
+	return nil, nil
+
 }
 
 func (h *RequestsHandler) readAdmissionReview(r *http.Request) (*admission.AdmissionReview, int, error) {
@@ -179,7 +182,7 @@ func (h *RequestsHandler) readAdmissionReview(r *http.Request) (*admission.Admis
 	return review, http.StatusOK, nil
 }
 
-func (h *RequestsHandler) lookupGenerator(namespace string, pod *corev1.Pod) (*inject.PatchGenerator, error) {
+func (h *RequestsHandler) lookupPod(namespace string, pod *corev1.Pod) (*inject.PatchGenerator, error) {
 	namespaceObj, err := h.clientset.CoreV1().Namespaces().Get(context.TODO(), namespace, metav1.GetOptions{})
 	if err != nil {
 		return nil, fmt.Errorf("failed to lookup pod's namespace (%s/%s): %v", namespace, pod.Name, err)
@@ -232,14 +235,59 @@ func (h *RequestsHandler) lookupGenerator(namespace string, pod *corev1.Pod) (*i
 	}, nil
 }
 
-func (h *RequestsHandler) handleAdmissionRequest(req *admission.AdmissionRequest) (k8tz.Patches, error) {
+func (h *RequestsHandler) lookupCronJob(namespace string, cronJob *batchv1.CronJob) (*inject.PatchGenerator, error) {
+	namespaceObj, err := h.clientset.CoreV1().Namespaces().Get(context.TODO(), namespace, metav1.GetOptions{})
+	if err != nil {
+		return nil, fmt.Errorf("failed to lookup cronJob's namespace (%s/%s): %v", namespace, cronJob.Name, err)
+	}
+
+	if _, ok := cronJob.Annotations[k8tz.InjectedAnnotation]; ok {
+		infoLogger.Printf("skipping cronJob (%s/%s) because its already injected", namespace, cronJob.Name)
+		return nil, nil
+	}
+
+	if val, ok := cronJob.Annotations[k8tz.InjectAnnotation]; ok {
+		if val == "false" {
+			infoLogger.Printf("skipping cronJob (%s/%s) because annotation on cronJob is explicitly false for injection", namespace, cronJob.Name)
+			return nil, nil
+		}
+	} else if val, ok := namespaceObj.Annotations[k8tz.InjectAnnotation]; ok {
+		if val == "false" {
+			infoLogger.Printf("skipping cronJob (%s/%s) because annotation on namespace is explicitly false for injection", namespace, cronJob.Name)
+			return nil, nil
+		}
+	} else if !h.InjectByDefault {
+		infoLogger.Printf("skipping cronJob (%s/%s) because no other instruction and injection disabled by default", namespace, cronJob.Name)
+		return nil, nil
+	}
+
+	timezone := h.DefaultTimezone
+	if val, ok := cronJob.Annotations[k8tz.TimezoneAnnotation]; ok {
+		timezone = val
+		infoLogger.Printf("explicit timezone requested on cronJob's (%s/%s) annotation: %s", namespace, cronJob.Name, val)
+	} else if val, ok := namespaceObj.Annotations[k8tz.TimezoneAnnotation]; ok {
+		timezone = val
+		infoLogger.Printf("explicit timezone requested on namespace (%s/%s) annotation: %s", namespace, cronJob.Name, val)
+	}
+
+	return &inject.PatchGenerator{
+		Strategy:           h.DefaultInjectionStrategy,
+		Timezone:           timezone,
+		InitContainerImage: h.BootstrapImage,
+		HostPathPrefix:     h.HostPathPrefix,
+		LocalTimePath:      h.LocalTimePath,
+		CronJobTimeZone:    h.CronJobTimeZone,
+	}, nil
+}
+
+func (h *RequestsHandler) handlePodAdmissionRequest(req *admission.AdmissionRequest) (k8tz.Patches, error) {
 	raw := req.Object.Raw
 	pod := corev1.Pod{}
 	if _, _, err := k8sdecode.Decode(raw, nil, &pod); err != nil {
 		return nil, fmt.Errorf("could not deserialize pod object: %v", err)
 	}
 
-	generator, err := h.lookupGenerator(req.Namespace, &pod)
+	generator, err := h.lookupPod(req.Namespace, &pod)
 	if err != nil {
 		return nil, fmt.Errorf("failed to lookup generator, error: %w", err)
 	}
@@ -247,6 +295,29 @@ func (h *RequestsHandler) handleAdmissionRequest(req *admission.AdmissionRequest
 	var patches k8tz.Patches
 	if generator != nil {
 		patches, err = generator.Generate(&pod, "")
+		if err != nil {
+			return nil, fmt.Errorf("failed to generate patches for pod, error: %w", err)
+		}
+	}
+
+	return patches, err
+}
+
+func (h *RequestsHandler) handleCronJobAdmissionRequest(req *admission.AdmissionRequest) (k8tz.Patches, error) {
+	raw := req.Object.Raw
+	cronjob := batchv1.CronJob{}
+	if _, _, err := k8sdecode.Decode(raw, nil, &cronjob); err != nil {
+		return nil, fmt.Errorf("could not deserialize pod object: %v", err)
+	}
+
+	generator, err := h.lookupCronJob(req.Namespace, &cronjob)
+	if err != nil {
+		return nil, fmt.Errorf("failed to lookup generator, error: %w", err)
+	}
+
+	var patches k8tz.Patches
+	if generator != nil {
+		patches, err = generator.Generate(&cronjob, "")
 		if err != nil {
 			return nil, fmt.Errorf("failed to generate patches for pod, error: %w", err)
 		}